requirements.txt
# =============================================================================
# MRI Signal Prediction — pip requirements
#
# Python: 3.10  (see environment.yml)
# GPU:    NVIDIA CUDA 12.6  (CPU-only users: drop the torch cu126 line below
#         and let `pip install torch torchvision` pick the default build.)
#
# ---------------------------------------------------------------------------
# Install order
# ---------------------------------------------------------------------------
# 1. Create the conda env:
#        conda env create -f environment.yml
#        conda activate mri_journal
#
# 2. Install the CUDA-matched PyTorch build from the PyTorch index:
#        pip install --index-url https://download.pytorch.org/whl/cu126 \
#            torch==2.8.0 torchvision==0.23.0
#
# 3. Install everything else:
#        pip install -r requirements.txt
#
# =============================================================================

# Deep-learning core
monai==1.5.0
einops==0.8.1
huggingface-hub==0.34.4
lpips==0.1.4
pytorch-ignite==0.4.11
safetensors==0.6.2
tensorboard==2.20.0
tensorboardX==2.6.4
torchio==0.20.21

# Scientific stack
numpy==2.2.6
scipy==1.15.3
pandas==2.3.2
scikit-image==0.25.2
scikit-learn==1.7.1
matplotlib==3.10.5
seaborn==0.13.2

# Medical imaging I/O
SimpleITK==2.5.2
itk==5.4.4.post1
nibabel==5.3.2
pydicom==3.0.1
tifffile==2025.5.10
imageio==2.37.0
pynrrd==1.1.3

# Utilities
tqdm==4.67.1
psutil==7.0.0
memory-profiler==0.61.0
Pillow==11.3.0
PyYAML==6.0.2
h5py==3.14.0
joblib==1.5.1
zarr==2.18.3
numcodecs==0.13.1

# Experiment tracking / optimisation (optional but imported by scripts)
clearml==2.0.3rc0
mlflow==3.3.1
optuna==4.5.0

# 3D rendering (only needed by scripts/visualization/visualise_samples.py).
# Install manually if you want to render vascular structures:
#     pip install vtk
# vtk==9.3.1

# Notebook support (for temp notebooks — optional)
# ipykernel==6.29.2
# ipython==8.21.0
# jupyterlab_widgets==3.0.10
